requirements.txt
python-dotenv==1.0.1
eth-account==0.13.4
web3==7.6.0
ccxt>=4.3.0
matplotlib>=3.7.0
rich>=13.0.0
